Update rest of toggles
